# backend-api/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/login/signin": {
      "post": {
        "operationId": "LoginController_logIn",
        "summary": "사용자 로그인",
        "description": "이메일과 비밀번호로 로그인하여 JWT 액세스 토큰을 발급받습니다.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "JWT 액세스 토큰",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (유효성 검사 실패)"
          },
          "401": {
            "description": "인증 실패 (이메일 또는 비밀번호 오류)"
          }
        },
        "tags": [
          "인증"
        ]
      }
    },
    "/v1/login/signup": {
      "post": {
        "operationId": "LoginController_create",
        "summary": "회원가입",
        "description": "새로운 사용자 계정을 생성합니다.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "회원가입 성공",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "userId": {
                      "type": "string"
                    },
                    "userEmail": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (유효성 검사 실패 또는 중복된 이메일)"
          }
        },
        "tags": [
          "인증"
        ]
      }
    },
    "/v1/login/signout": {
      "post": {
        "operationId": "LoginController_signOut",
        "summary": "로그아웃",
        "description": "현재 세션을 종료하고 리프레시 토큰을 삭제합니다.",
        "parameters": [],
        "responses": {
          "201": {
            "description": "로그아웃 성공",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        },
        "tags": [
          "인증"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "VanillaMeta API Documentation",
    "description": "VanillaMeta는 기업용 비즈니스 인텔리전스(BI) 웹 애플리케이션으로, 사용자가 코드 작성 없이 다양한 데이터베이스에 연결하여 시각화를 생성하고 대시보드를 구축할 수 있습니다.\n\n## 주요 기능\n- 다중 데이터베이스 연결 지원\n- 50개 이상의 차트 타입 제공\n- 대시보드 생성 및 공유\n- 실시간 데이터 시각화\n\n## 인증\nJWT Bearer 토큰을 사용합니다. 로그인 API를 통해 액세스 토큰을 발급받은 후 Authorization 헤더에 Bearer {token} 형식으로 전달하세요.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/v1",
      "description": "로컬 개발 서버"
    },
    {
      "url": "https://dev-api.vanillameta.com/v1",
      "description": "개발 서버"
    },
    {
      "url": "https://api.vanillameta.com/v1",
      "description": "프로덕션 서버"
    }
  ],
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT 액세스 토큰을 입력하세요",
        "in": "header"
      }
    },
    "schemas": {
...
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "사용자 이메일 주소",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "사용자 비밀번호",
            "example": "password123"
          }
        },
        "required": [
          "userId",
          "password"
        ]
      },
      "CreateLoginDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "사용자 ID (고유 식별자)",
            "example": "johndoe",
            "minLength": 3,
            "maxLength": 50
          },
          "password": {
            "type": "string",
            "description": "사용자 비밀번호",
            "example": "StrongPassword123!",
            "minLength": 8
          },
          "email": {
            "type": "string",
            "description": "사용자 이메일 주소",
            "example": "johndoe@example.com",
            "format": "email"
          },
          "applicationNote": {
            "type": "string",
            "description": "가입 신청 메모 (선택사항)",
            "example": "데이터 분석팀에서 VanillaMeta를 사용하고자 합니다."
          }
        },
        "required": [
          "userId",
          "password",
          "email"
        ]
      },
...
    },
...
  },
  "security": [
    {
      "AccessToken": []
    }
  ]
}